* commit '52495852fe128148f565f05f0afa939c9c1b7527':
  Remove references to Roadmap in HomePage functional test and FlowSpecs
  Delete Roadmap functional test files
  Remove Roadmap from core routes
  Remove Roadmap link from footer
  Delete Roadmap page
  Delete Roadmap link from HomePage element definitions
  BA-1102 - Fix incorrect deadline in email notification
  BA-1104 - Fix issue where default user avatar is deleted on server

+++ modules/core/server/controllers/CoreServerHelpers.ts
@@ -125,7 +125,7 @@ class CoreServerHelpers {
 			},
 			deleteOldImage: () => {
 				return new Promise((resolve, reject) => {
-					if (existingImageUrl !== doc[field]) {
+					if (existingImageUrl !== doc[field] && existingImageUrl !== 'img/default.png') {
 						fs.unlink('./public/' + existingImageUrl, () => {
 							resolve();
 						});
